config.py
from decouple import config

## Enter your Open API Key here
OPENAI_API_KEY = config('OPENAI_API_KEY')

class Config(object):
    DEBUG = True
    TESTING = False

class DevelopmentConfig(Config):
    SECRET_KEY = "this-is-a-super-secret-key"

config = {
    'development': DevelopmentConfig,
    'testing': DevelopmentConfig,
    'production': DevelopmentConfig
}

